requirements.txt
py-trello==0.3.1
python-taiga==0.4.0

